feat: defaut expand the empty root node to show the placeholder
